+++ src/PubSub/AbstractEventSubscriber.php
@@ -1,6 +1,29 @@
 <?php
 /**
- * @author hollodotme
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
+declare(strict_types = 1);
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
  */
 
 namespace IceHawk\IceHawk\PubSub;
@@ -32,13 +55,13 @@ abstract class AbstractEventSubscriber implements SubscribesToEvents
 		$namespaceComponents = explode( "\\", get_class( $event ) );
 		$methodName          = sprintf( 'when%s', preg_replace( "#Event$#", '', end( $namespaceComponents ) ) );
 
-		if ( is_callable( [ $this, $methodName ] ) )
+		if ( is_callable( [$this, $methodName] ) )
 		{
 			$this->{$methodName}( $event );
 		}
 		else
 		{
-			throw ( new EventSubscriberMethodNotCallable() )->withMethodName( $methodName );
+			throw (new EventSubscriberMethodNotCallable())->withMethodName( $methodName );
 		}
 	}
 }