+++ src/RequestHandlers/OptionsRequestHandler.php
@@ -1,6 +1,29 @@
 <?php
 /**
- * @author h.woltersdorf
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
+declare(strict_types = 1);
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
  */
 
 namespace IceHawk\IceHawk\RequestHandlers;
@@ -20,14 +43,14 @@ final class OptionsRequestHandler extends AbstractRequestHandler
 	{
 		$allowedRequestMethods = $this->getOptions();
 
-		( new Options( $allowedRequestMethods ) )->respond();
+		(new Options( $allowedRequestMethods ))->respond();
 	}
 
 	private function getOptions() : array
 	{
 		$handlerRoutes = $this->getHandlerRoutes();
 
-		$requestMethods = [ ];
+		$requestMethods = [];
 		foreach ( $handlerRoutes as $handlerRoute )
 		{
 			$handler        = $handlerRoute->getRequestHandler();
@@ -42,7 +65,7 @@ final class OptionsRequestHandler extends AbstractRequestHandler
 		$readRoutes  = $this->config->getReadRoutes();
 		$writeRoutes = $this->config->getWriteRoutes();
 		$routes      = array_merge( $readRoutes, $writeRoutes );
-		
+
 		$requestInfo = $this->config->getRequestInfo();
 		$router      = new OptionsRouter( $routes );
 
@@ -51,7 +74,7 @@ final class OptionsRequestHandler extends AbstractRequestHandler
 
 	private function getImplementedRequestMethods( HandlesRequest $handler ) : array
 	{
-		$requestMethods = [ ];
+		$requestMethods = [];
 
 		foreach ( HandlerMethodInterfaceMap::HTTP_METHODS as $requestMethod => $interface )
 		{