+++ tests/Unit/IceHawkTest.php
@@ -1,4 +1,31 @@
 <?php
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
+declare(strict_types = 1);
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
 namespace IceHawk\IceHawk\Tests\Unit\Defaults;
 
 use IceHawk\IceHawk\Constants\HttpCode;
@@ -56,7 +83,7 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		$initializedEvent  = new IceHawkWasInitializedEvent( $requestInfo );
 
 		$eventListener = $this->getMockBuilder( SubscribesToEvents::class )
-		                      ->setMethods( [ 'acceptsEvent', 'notify' ] )
+		                      ->setMethods( ['acceptsEvent', 'notify'] )
 		                      ->getMockForAbstractClass();
 
 		$eventListener->expects( $this->at( 0 ) )
@@ -82,7 +109,7 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
 		$config->expects( $this->once() )
 		       ->method( 'getEventSubscribers' )
-		       ->willReturn( [ $eventListener ] );
+		       ->willReturn( [$eventListener] );
 
 		$iceHawk = new IceHawk( $config, new IceHawkDelegate() );
 		$iceHawk->init();
@@ -131,8 +158,8 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		$route = new ReadRoute( new Literal( '/test' ), $requestHandler );
 
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
-		$config->expects( $this->once() )->method( 'getReadRoutes' )->willReturn( [ $route ] );
-		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [ ] );
+		$config->expects( $this->once() )->method( 'getReadRoutes' )->willReturn( [$route] );
+		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [] );
 
 		$delegate = new IceHawkDelegate();
 
@@ -169,8 +196,8 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		$route = new WriteRoute( new Literal( '/test' ), $requestHandler );
 
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
-		$config->expects( $this->once() )->method( 'getWriteRoutes' )->willReturn( [ $route ] );
-		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [ ] );
+		$config->expects( $this->once() )->method( 'getWriteRoutes' )->willReturn( [$route] );
+		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [] );
 
 		$delegate = new IceHawkDelegate();
 
@@ -197,7 +224,7 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 					new WriteRoute( new Literal( '/do/whatever/you/want' ), new DeleteRequestHandler() ),
 				],
 				'/do/this',
-				[ 'POST' ],
+				['POST'],
 			],
 		];
 	}
@@ -237,7 +264,7 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 	public function testPublishesEventsWhenHandlingReadRequest()
 	{
 		$config      = $this->getMockBuilder( ConfiguresIceHawk::class )->getMockForAbstractClass();
-		$requestInfo = new RequestInfo( [ 'REQUEST_METHOD' => 'GET', 'REQUEST_URI' => '/test', ] );
+		$requestInfo = new RequestInfo( ['REQUEST_METHOD' => 'GET', 'REQUEST_URI' => '/test',] );
 
 		$requestHandler = $this->getMockBuilder( HandlesGetRequest::class )->getMockForAbstractClass();
 		$requestHandler->expects( $this->once() )->method( 'handle' );
@@ -251,31 +278,31 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		$handledEvent      = new ReadRequestWasHandledEvent( $request );
 
 		$eventListener = $this->getMockBuilder( SubscribesToEvents::class )
-		                      ->setMethods( [ 'acceptsEvent', 'notify' ] )
+		                      ->setMethods( ['acceptsEvent', 'notify'] )
 		                      ->getMockForAbstractClass();
 
 		$eventListener->expects( $this->exactly( 4 ) )
 		              ->method( 'acceptsEvent' )
 		              ->withConsecutive(
-			              [ $this->equalTo( $initializingEvent ) ],
-			              [ $this->equalTo( $initEvent ) ],
-			              [ $this->equalTo( $handlingEvent ) ],
-			              [ $this->equalTo( $handledEvent ) ]
+			              [$this->equalTo( $initializingEvent )],
+			              [$this->equalTo( $initEvent )],
+			              [$this->equalTo( $handlingEvent )],
+			              [$this->equalTo( $handledEvent )]
 		              )
 		              ->willReturn( true );
 
 		$eventListener->expects( $this->exactly( 4 ) )
 		              ->method( 'notify' )
 		              ->withConsecutive(
-			              [ $this->equalTo( $initializingEvent ) ],
-			              [ $this->equalTo( $initEvent ) ],
-			              [ $this->equalTo( $handlingEvent ) ],
-			              [ $this->equalTo( $handledEvent ) ]
+			              [$this->equalTo( $initializingEvent )],
+			              [$this->equalTo( $initEvent )],
+			              [$this->equalTo( $handlingEvent )],
+			              [$this->equalTo( $handledEvent )]
 		              );
 
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
-		$config->expects( $this->once() )->method( 'getReadRoutes' )->willReturn( [ $route ] );
-		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [ $eventListener ] );
+		$config->expects( $this->once() )->method( 'getReadRoutes' )->willReturn( [$route] );
+		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [$eventListener] );
 
 		$delegate = new IceHawkDelegate();
 
@@ -290,7 +317,7 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 	public function testPublishesEventsWhenHandlingWriteRequest()
 	{
 		$config      = $this->getMockBuilder( ConfiguresIceHawk::class )->getMockForAbstractClass();
-		$requestInfo = new RequestInfo( [ 'REQUEST_METHOD' => 'POST', 'REQUEST_URI' => '/test', ] );
+		$requestInfo = new RequestInfo( ['REQUEST_METHOD' => 'POST', 'REQUEST_URI' => '/test',] );
 
 		$requestHandler = $this->getMockBuilder( HandlesPostRequest::class )->getMockForAbstractClass();
 		$requestHandler->expects( $this->once() )->method( 'handle' );
@@ -304,31 +331,31 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		$handledEvent      = new WriteRequestWasHandledEvent( $request );
 
 		$eventListener = $this->getMockBuilder( SubscribesToEvents::class )
-		                      ->setMethods( [ 'acceptsEvent', 'notify' ] )
+		                      ->setMethods( ['acceptsEvent', 'notify'] )
 		                      ->getMockForAbstractClass();
 
 		$eventListener->expects( $this->exactly( 4 ) )
 		              ->method( 'acceptsEvent' )
 		              ->withConsecutive(
-			              [ $this->equalTo( $initializingEvent ) ],
-			              [ $this->equalTo( $initEvent ) ],
-			              [ $this->equalTo( $handlingEvent ) ],
-			              [ $this->equalTo( $handledEvent ) ]
+			              [$this->equalTo( $initializingEvent )],
+			              [$this->equalTo( $initEvent )],
+			              [$this->equalTo( $handlingEvent )],
+			              [$this->equalTo( $handledEvent )]
 		              )
 		              ->willReturn( true );
 
 		$eventListener->expects( $this->exactly( 4 ) )
 		              ->method( 'notify' )
 		              ->withConsecutive(
-			              [ $this->equalTo( $initializingEvent ) ],
-			              [ $this->equalTo( $initEvent ) ],
-			              [ $this->equalTo( $handlingEvent ) ],
-			              [ $this->equalTo( $handledEvent ) ]
+			              [$this->equalTo( $initializingEvent )],
+			              [$this->equalTo( $initEvent )],
+			              [$this->equalTo( $handlingEvent )],
+			              [$this->equalTo( $handledEvent )]
 		              );
 
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
-		$config->expects( $this->once() )->method( 'getWriteRoutes' )->willReturn( [ $route ] );
-		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [ $eventListener ] );
+		$config->expects( $this->once() )->method( 'getWriteRoutes' )->willReturn( [$route] );
+		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [$eventListener] );
 
 		$delegate = new IceHawkDelegate();
 
@@ -363,9 +390,9 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		               );
 
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
-		$config->expects( $this->once() )->method( 'getReadRoutes' )->willReturn( [ ] );
+		$config->expects( $this->once() )->method( 'getReadRoutes' )->willReturn( [] );
 		$config->method( 'getFinalReadResponder' )->willReturn( $finalResponder );
-		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [ ] );
+		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [] );
 
 		$delegate = new IceHawkDelegate();
 
@@ -402,9 +429,9 @@ class IceHawkTest extends \PHPUnit_Framework_TestCase
 		               );
 
 		$config->expects( $this->once() )->method( 'getRequestInfo' )->willReturn( $requestInfo );
-		$config->expects( $this->once() )->method( 'getWriteRoutes' )->willReturn( [ ] );
+		$config->expects( $this->once() )->method( 'getWriteRoutes' )->willReturn( [] );
 		$config->method( 'getFinalWriteResponder' )->willReturn( $finalResponder );
-		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [ ] );
+		$config->expects( $this->once() )->method( 'getEventSubscribers' )->willReturn( [] );
 
 		$delegate = new IceHawkDelegate();