+++ tests/Unit/PubSub/EventSubscriberTest.php
@@ -1,4 +1,31 @@
 <?php
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
+declare(strict_types = 1);
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
 namespace IceHawk\IceHawk\Tests\Unit\Fixtures\PubSub;
 
 use IceHawk\IceHawk\Defaults\RequestInfo;
@@ -22,12 +49,12 @@ class EventSubscriberTest extends \PHPUnit_Framework_TestCase
 			new HandlingReadRequestEvent( new ReadRequest( RequestInfo::fromEnv(), new ReadRequestInput( [] ) ) );
 
 		$mock = $this->getMockBuilder( TestEventSubscriber::class )
-		             ->setMethods( [ 'getAcceptedEvents', 'whenIceHawkWasInitialized' ] )
+		             ->setMethods( ['getAcceptedEvents', 'whenIceHawkWasInitialized'] )
 		             ->getMockForAbstractClass();
 
 		$mock->expects( $this->exactly( 2 ) )
 		     ->method( 'getAcceptedEvents' )
-		     ->willReturn( [ IceHawkWasInitializedEvent::class ] );
+		     ->willReturn( [IceHawkWasInitializedEvent::class] );
 
 		/** @var TestEventSubscriber $mock */
 		$this->assertTrue( $mock->acceptsEvent( $initEvent ) );
@@ -37,25 +64,25 @@ class EventSubscriberTest extends \PHPUnit_Framework_TestCase
 	public function testNotImplementedListenerMethodThrowsException()
 	{
 		$initEvent = new IceHawkWasInitializedEvent( RequestInfo::fromEnv() );
-	
-		try 
+
+		try
 		{
 			$mock = $this->getMockBuilder( TestEventSubscriber::class )
-			             ->setMethods( [ 'getAcceptedEvents' ] )
+			             ->setMethods( ['getAcceptedEvents'] )
 			             ->getMockForAbstractClass();
 
 			/** @var TestEventSubscriber $mock */
 			$mock->notify( $initEvent );
-			
-			$this->fail('No Exception thrown');
-		} 
+
+			$this->fail( 'No Exception thrown' );
+		}
 		catch ( EventSubscriberMethodNotCallable $ex )
 		{
 			$this->assertEquals( 'whenIceHawkWasInitialized', $ex->getMethodName() );
 		}
 		catch ( \Throwable $throwable )
 		{
-			$this->fail( 'Wrong exception thrown');
+			$this->fail( 'Wrong exception thrown' );
 		}
 	}
 
@@ -64,7 +91,7 @@ class EventSubscriberTest extends \PHPUnit_Framework_TestCase
 		$initEvent = new IceHawkWasInitializedEvent( RequestInfo::fromEnv() );
 
 		$mock = $this->getMockBuilder( TestEventSubscriber::class )
-		             ->setMethods( [ 'getAcceptedEvents', 'whenIceHawkWasInitialized' ] )
+		             ->setMethods( ['getAcceptedEvents', 'whenIceHawkWasInitialized'] )
 		             ->getMockForAbstractClass();
 
 		$mock->expects( $this->once() )->method( 'whenIceHawkWasInitialized' )->with( $initEvent );