+++ tests/Unit/RequestHandlers/OptionsRequestHandlerTest.php
@@ -1,4 +1,31 @@
 <?php
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
+declare(strict_types = 1);
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
 namespace IceHawk\IceHawk\Tests\Unit\RequestHandlers;
 
 use IceHawk\IceHawk\Defaults\RequestInfo;
@@ -20,7 +47,7 @@ class OptionsRequestHandlerTest extends \PHPUnit_Framework_TestCase
 	 * @dataProvider RouteProvider
 	 * @runInSeparateProcess
 	 */
-	public function testHeaderOutput( array $readRoutes, array $writeRoutes, string $uri, array $expectedMethods)
+	public function testHeaderOutput( array $readRoutes, array $writeRoutes, string $uri, array $expectedMethods )
 	{
 		$config      = $this->getMockBuilder( ConfiguresIceHawk::class )->getMockForAbstractClass();
 		$requestInfo = new RequestInfo(
@@ -38,7 +65,7 @@ class OptionsRequestHandlerTest extends \PHPUnit_Framework_TestCase
 		$optionsRequestHandler->handleRequest();
 
 		$expectedHeader = 'Allow: ' . join( ',', $expectedMethods );
-		
+
 		$this->assertContains( $expectedHeader, xdebug_get_headers() );
 	}
 
@@ -58,7 +85,7 @@ class OptionsRequestHandlerTest extends \PHPUnit_Framework_TestCase
 					new WriteRoute( new Literal( '/do/whatever/you/want' ), new DeleteRequestHandler() ),
 				],
 				'/do/this',
-				[ 'POST' ],
+				['POST'],
 			],
 			[
 				[
@@ -73,7 +100,7 @@ class OptionsRequestHandlerTest extends \PHPUnit_Framework_TestCase
 					new WriteRoute( new Literal( '/this' ), new DeleteRequestHandler() ),
 				],
 				'/this',
-				[ 'HEAD', 'GET', 'HEAD', 'POST', 'PUT', 'DELETE' ],
+				['HEAD', 'GET', 'HEAD', 'POST', 'PUT', 'DELETE'],
 			],
 		];
 	}