+++ tests/Unit/Requests/ReadRequestInputTest.php
@@ -1,6 +1,29 @@
 <?php
 /**
- * @author h.woltersdorf
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
+declare(strict_types = 1);
+/**
+ * Copyright (c) 2016 Holger Woltersdorf & Contributors
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
  */
 
 namespace IceHawk\IceHawk\Tests\Unit\Requests;
@@ -13,19 +36,19 @@ class ReadRequestInputTest extends \PHPUnit_Framework_TestCase
 	{
 		return [
 			[
-				[ 'unit' => 'test', 'test' => 'unit' ],
+				['unit' => 'test', 'test' => 'unit'],
 				'unit',
 				'test',
 			],
 			[
-				[ 'unit' => 'test', 'test' => 'unit' ],
+				['unit' => 'test', 'test' => 'unit'],
 				'test',
 				'unit',
 			],
 			[
-				[ 'unit' => [ 'test' => 'unit' ] ],
+				['unit' => ['test' => 'unit']],
 				'unit',
-				[ 'test' => 'unit' ],
+				['test' => 'unit'],
 			],
 		];
 	}
@@ -54,15 +77,15 @@ class ReadRequestInputTest extends \PHPUnit_Framework_TestCase
 	{
 		return [
 			[
-				[ 'unit' => 'test', 'test' => 'unit' ],
+				['unit' => 'test', 'test' => 'unit'],
 				'blubb',
 			],
 			[
-				[ 'unit' => 'test', 'test' => 'unit' ],
+				['unit' => 'test', 'test' => 'unit'],
 				'blubb',
 			],
 			[
-				[ 'unit' => [ 'test' => 'unit' ] ],
+				['unit' => ['test' => 'unit']],
 				'blubb',
 			],
 		];